Added a filter image to server.ts file
