+++ README.md
@@ -1,4 +1,3 @@
-# Project_Udacity
 # Full Stack Apps on AWS Project
 
 You have been hired as a software engineer to develop an application that will help the FBI find missing people.  The application will upload images to the FBI cloud database hosted in AWS. This will allow the FBI to run facial recognition software on the images to detect a match. You will be developing a NodeJS server and deploying it on AWS Elastic Beanstalk.
