+++ package.json
@@ -1,18 +1,35 @@
 {
-  "name": "udacity-project-2",
+  "name": "udacity-c2-image-filter",
   "version": "1.0.0",
   "description": "",
-  "main": "server.js",
-  "type": "module",
+  "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node server.js"
+    "clean": "rimraf www/ || true",
+    "build": "npm run clean && tsc && cp package.json www/package.json && mkdir www/tmp/ && cd www && zip -r Archive.zip . && cd ..",
+    "dev": "ts-node-dev --respawn --transpile-only ./src/server.ts"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/grutt/udacity-c2-image-filter.git"
   },
   "author": "",
   "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/grutt/udacity-c2-image-filter/issues"
+  },
+  "homepage": "https://github.com/grutt/udacity-c2-image-filter#readme",
+  "devDependencies": {
+    "@types/express": "^4.17.0",
+    "@types/node": "^11.13.17",
+    "ts-node-dev": "^1.0.0-pre.40",
+    "tslint": "^5.18.0",
+    "typescript": "^3.5.3"
+  },
   "dependencies": {
-    "body-parser": "^1.20.1",
-    "express": "^4.18.2",
-    "jimp": "^0.16.2"
+    "express": "^4.17.1",
+    "jimp": "^0.16.1",
+    "rimraf": "^3.0.2",
+    "lodash": "^4.17.15"
   }
 }
