+++ tslint.json
@@ -1,0 +1,10 @@
+{
+    // "defaultSeverity": "error",
+    // "extends": [
+    //     "tslint:recommended"
+    // ],
+    // "jsRules": {},
+    // "rules": {
+    // },
+    // "rulesDirectory": []
+}
